demos/demo_S3_UniversalAttack/evaluate.py
"""
Evaluate a universal adversarial image against target MLLMs.

Tests the attack success rate: does the model respond with unsafe content
when shown the adversarial image alongside diverse questions?

Usage:
    python evaluate.py --image outputs/universal_final.png
    python evaluate.py --image outputs/universal_final.png --target-models qwen2_5_vl_3b
"""

import argparse
import json
import os
import sys

import torch
from PIL import Image
from torchvision import transforms

sys.path.insert(0, os.path.dirname(__file__))
from config import ATTACK_CONFIG
from dataset import AttackDataset


def load_image_tensor(path: str, size: tuple = (448, 448)) -> torch.Tensor:
    """Load image as (1, 3, H, W) tensor in [0, 1]."""
    img = Image.open(path).convert("RGB")
    transform = transforms.Compose([
        transforms.Resize(size),
        transforms.ToTensor(),
    ])
    return transform(img).unsqueeze(0)


def main(args):
    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

    # Load adversarial image
    adv_image = load_image_tensor(args.image, tuple(args.image_size)).to(device)
    print(f"Adversarial image: {args.image}")
    print(f"Image size: {adv_image.shape}")

    dataset = AttackDataset()

    # Evaluate against each model
    for model_key in args.target_models:
        print(f"\n{'='*60}")
        print(f"Evaluating: {model_key}")
        print(f"{'='*60}")

        from attack import get_wrapper_for_model
        wrapper = get_wrapper_for_model(model_key, device)

        results = []
        success_count = 0
        total_count = 0

        # Test with adversarial questions
        test_questions = (
            dataset.adversarial[:args.num_adversarial]
            + dataset.safe[:args.num_safe]
        )

        for i, question in enumerate(test_questions):
            is_adversarial = i < args.num_adversarial
            response = wrapper.generate(adv_image, question, max_new_tokens=150)

            # Simple ASR check: does response start with target-like affirmation?
            target_lower = args.target_phrase.lower()
            response_lower = response.strip().lower()
            is_success = (
                response_lower.startswith(target_lower[:10])
                or target_lower in response_lower[:50]
            )

            results.append({
                "question": question,
                "response": response,
                "is_adversarial": is_adversarial,
                "attack_success": is_success,
            })

            if is_adversarial:
                total_count += 1
                if is_success:
                    success_count += 1

            tag = "ADV" if is_adversarial else "SAFE"
            status = "HIT" if is_success else "MISS"
            print(f"  [{tag}][{status}] Q: {question[:60]}...")
            print(f"          A: {response[:100]}...")

        asr = success_count / max(total_count, 1) * 100
        print(f"\n  Attack Success Rate (adversarial): {asr:.1f}% "
              f"({success_count}/{total_count})")

        # Save results
        os.makedirs(args.output_dir, exist_ok=True)
        result_path = os.path.join(args.output_dir, f"eval_{model_key}.json")
        with open(result_path, "w") as f:
            json.dump({
                "model": model_key,
                "image": args.image,
                "target_phrase": args.target_phrase,
                "asr_adversarial": asr,
                "total_adversarial": total_count,
                "success_adversarial": success_count,
                "results": results,
            }, f, indent=2, ensure_ascii=False)
        print(f"  Results saved: {result_path}")

        wrapper.unload()


if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Evaluate Universal Attack")
    parser.add_argument("--image", type=str, required=True,
                        help="Path to universal adversarial image")
    parser.add_argument("--target-models", nargs="+",
                        default=ATTACK_CONFIG["target_models"])
    parser.add_argument("--target-phrase", type=str,
                        default=ATTACK_CONFIG["target_phrase"])
    parser.add_argument("--image-size", type=int, nargs=2,
                        default=list(ATTACK_CONFIG["image_size"]))
    parser.add_argument("--num-adversarial", type=int, default=20)
    parser.add_argument("--num-safe", type=int, default=10)
    parser.add_argument("--output-dir", type=str, default="outputs/results")
    main(parser.parse_args())
