demos/demo_S3_UniversalAttack/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
accelerate>=0.25.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
qwen-vl-utils>=0.0.2
